backend/requirements.txt
flask
pymongo
flask_bcrypt
flask_jwt_extended
requests
openai
python-dotenv
flask-cors